Updated the configuration of the Prefab Processing Stack.
Changed the Prefab Processing Stack's configuration to use the new ConfigurableStack. This makes it easier to inject additional nodes into the stack, which is used to update the config for the multiplayer gem.
This update also adds some trace context entries, making it easier now to see what configuration was used when processing and what processor caused the issue.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>